Corrections optimisations

+++ mysqlbackup/mysql-backup.sh
@@ -1,9 +1,17 @@
 #!/bin/bash
- 
+
+## Inspiré d'un script trouvé sur http://phpnews.fr 
+## Et déjà modifié par http://www.mercereau.info et les commentaires de http://e-concept-applications.fr
+## 2013 | Yvan GODARD | http://www.yvangodard.me | godardyvan@gmail.com
+
 ## Variables
+# Utilisateur SQL à activer si besoin
+# USER='root'
+# PASS='xxxxxxxxx'
 # Le script
 CURRENT_DIR=$(dirname $0)
 SCRIPT_NAME=$(basename $0)
+HOSTNAME=$(hostname)
 # Répertoire de sauvergarde des dump SQL
 LOCATION="/home/mysqldump"
 # Nom du backup
@@ -11,15 +19,14 @@ DATANAME="databasebackup-$(date +%d.%m.%y@%Hh%M)"
 # Répertoire temporaire
 DATATMP="$LOCATION/temp"
 # Mail pour l'envoi du rapport
-MAIL_ADMIN="admin@reseauenscene.fr"
+MAIL_ADMIN="monmail@moi.me"
 # Bases SQL à exclure
 EXCLUSIONS='(information_schema)'
 # Version du script
-SCR_VERS="1.2"
+SCR_VERS="1.3"
 # Les logs
 LOGSLOCATION="/var/log/mysql-backup"
 LOG_OUT="$LOGSLOCATION/out.log"
-LOG_ERROR="$LOGSLOCATION/error.log"
 LOG_CUMUL="$LOGSLOCATION/mysql-backup.log"
 DATE_DU_JOUR=$(date)
 # Initialisation signal erreur 
@@ -33,19 +40,28 @@ fi
 
 umask 027
 
-## Redirection des sorties vers nos logs
-mkdir -p $LOCATION
-mkdir -p $LOGSLOCATION
-[ -f $LOG_OUT ] && rm  $LOG_OUT
-[ -f $LOG_ERROR ] && rm  $LOG_ERROR
-exec 1>> $LOG_OUT
-exec 2>> $LOG_ERROR
- 
+## Redirection des sorties vers nos logs : création des dossiers nécessaires
+if [ ! -d $$LOGSLOCATION ]; then
+    mkdir -p $LOGSLOCATION
+    [ $? -ne 0 ] && ERROR=1 && echo "*** Problème pour créer le dossier $LOGSLOCATION ***" && echo "Il sera impossible de journaliser le processus."
+fi
+if [ ! -d $LOCATION ]; then
+    mkdir -p $LOCATION 
+    [ $? -ne 0 ] && echo "*** Problème pour créer le dossier $LOCATION ***" && echo "Il est impossible de poursuivre la sauvegarde." && exit 1
+fi
+
+# Suppression des anciens logs temporaires
+[ -f $LOG_OUT ] && rm $LOG_OUT
+
+# Ouverture de notre fichier de log 
 echo "" >> $LOG_OUT
 echo "****************************** $DATE_DU_JOUR ******************************" >> $LOG_OUT
 echo "" >> $LOG_OUT
+echo "Machine : " $HOSTNAME >> $LOG_OUT
+echo "" >> $LOG_OUT
  
 cd $LOCATION
+[ $? -ne 0 ] && echo "*** Problème pour accéder au dossier $LOCATION ***" && echo "Il est impossible de poursuivre la sauvegarde." && exit 1
  
 ## En fonction du jour, changement du nombre de backup à garder et du répertoire de destination
 if [ "$( date +%w )" == "0" ]; then
@@ -63,12 +79,15 @@ else
 fi
  
 ## Création d'un répertoire temporaire pour la sauvegarde avant de zipper l'ensemble des dumps
-mkdir -p "${DATATMP}/${DATANAME}"
-[ $? -ne 0 ] && ERROR=1 && echo "*** Problème pour créer le dossier ${DATATMP}/${DATANAME} ***"
+mkdir -p ${DATATMP}/${DATANAME}
+[ $? -ne 0 ] && ERROR=1 && echo "*** Problème pour créer le dossier ${DATATMP}/${DATANAME} ***" >> $LOG_OUT
  
 # On place dans un tableau le nom de toutes les bases de données du serveur
+# Version avec mot de passe
+# databases="$(mysql --user=$USER --password=$PASS -Bse 'show databases' | grep -v -E $EXCLUSIONS)"
+# Version sans mot de passe
 databases="$(mysql -Bse 'show databases' | grep -v -E $EXCLUSIONS)"
-[ $? -ne 0 ] && ERROR=1 && echo "*** Problème pour obtenir la liste des bases à dumper ***"
+[ $? -ne 0 ] && ERROR=1 && echo "*** Problème pour obtenir la liste des bases à dumper ***" >> $LOG_OUT
 echo "Bases de données à traiter :" >> $LOG_OUT
  
 # Sauvegarde de toutes les bases dans le fichier du jour
@@ -76,43 +95,44 @@ echo "Bases de données à traiter :" >> $LOG_OUT
 for database in ${databases[@]}
 do
     echo "- ${database}.sql"  >> $LOG_OUT
+    # Version avec mot de passe
+    # mysqldump  --user=$USER --password=$PASS --events --quick --add-locks --lock-tables --extended-insert $database  > ${DATATMP}/${DATANAME}/${database}.sql
+    # Version sans mot de passe
     mysqldump --events --quick --add-locks --lock-tables --extended-insert $database  > ${DATATMP}/${DATANAME}/${database}.sql
-    [ $? -ne 0 ] && ERROR=1
+    [ $? -ne 0 ] && ERROR=1 && echo "*** Problème sur le dump de la base $database ***"  >> $LOG_OUT
 done
  
 ## On commpresse (TAR) tous et on créé un lien symbolique pour le dernier
 cd ${DATATMP}
 echo "Création de l'archive ${DATADIR}/${DATANAME}.sql.gz" >> $LOG_OUT
 tar -czf ${DATADIR}/${DATANAME}.sql.gz ${DATANAME}
-[ $? -ne 0 ] && ERROR=1 && echo "*** Problème lors de la création de l'archive ${DATADIR}/${DATANAME}.sql.gz ***"
+[ $? -ne 0 ] && ERROR=1 && echo "*** Problème lors de la création de l'archive ${DATADIR}/${DATANAME}.sql.gz ***" >> $LOG_OUT
 cd ${DATADIR}
 chmod 600 ${DATANAME}.sql.gz
 [ -f last.sql.gz ] &&  rm last.sql.gz
-ln -s ${DATANAME}.sql.gz last.sql.gz
+ln -s ${DATADIR}/${DATANAME}.sql.gz ${DATADIR}/last.sql.gz
  
 ## On supprime le répertoire temporaire
-rm -rf ${DATATMP}/${DATANAME}
+ [ -d ${DATATMP}/${DATANAME} ] && rm -rf ${DATATMP}/${DATANAME}
  
 ## On supprime les anciens backups
 echo "Suppression des vieux DUMP éventuels" >> $LOG_OUT
-find $DATADIR -name "*.sql.gz" -mtime +$KEEP_NUMBER -print -exec rm {} \;
+find $DATADIR -name "*.sql.gz" -mtime +$KEEP_NUMBER -print -exec rm {} \; >> $LOG_OUT
 [ $? -ne 0 ] && ERROR=1
- 
-## On fusionne le log d'erreur dans le log de sortie standard si il y a eu une erreur
-[ $ERROR -ne 0 ] && echo "Log d'erreur :" >> $LOG_OUT ; cat $LOG_ERROR >> $LOG_OUT
  
 ## Envoi d'un email de notification
 if [ $ERROR -ne 0 ]
-then
-echo "Problème lors de l'éxécution de (${0}). Merci de corriger le processus." >> $LOG_OUT
-mail -s "[FAILED] Rapport Dump MySql (${0})" $MAIL_ADMIN <"${LOG_OUT}"
-else
-echo "Script de dump des bases MySql (${0}) exécuté avec succès."  >> $LOG_OUT
-mail -s "[OK] Rapport Dump MySql (${0})" $MAIL_ADMIN <"${LOG_OUT}"
+    then
+        echo "Problème lors de l'éxécution de (${0}). Merci de corriger le processus." >> $LOG_OUT
+        mail -s "[FAILED] Rapport Dump MySql (${0})" $MAIL_ADMIN <"${LOG_OUT}"
+    else
+        echo "Script de dump des bases MySql (${0}) exécuté avec succès."  >> $LOG_OUT
+        mail -s "[OK] Rapport Dump MySql (${0})" $MAIL_ADMIN <"${LOG_OUT}"
 fi
  
 cat $LOG_OUT >> $LOG_CUMUL
 [ -f $LOG_OUT ] && rm  $LOG_OUT
-[ -f $LOG_ERROR ] && rm  $LOG_ERROR
- 
+
+[ $ERROR -ne 0 ] && exit 1
+
 exit 0
